Create README.md
# Add_DOI_to_references

This short script was developed for Google Apps to take a list of journal article references that do not include the DOI number with them, and add the DOI to the end. The script uses the CrossRef API to look up the DOI based on key words from the reference (title, author, etc.). It does a pretty good job, though after you will still want to spot check the DOIs to see if they are going to the correct URL for the article.  The  script is handy for when you have lots of references without DOIs.  Enjoy.